Git annex (#136)
The main point of this `fix` commit is to tell git-annex about some files that @valosekj added in the `jv/update_T2w_labels` branch and uploaded to amazon. For some reason:

* the files *were* uploaded;
* but git-annex hadn't recorded that fact;
* but `git-annex copy --to` saw that the files existed on amazon;
* so it still didn't record the fact in the git-annex branch.

The second point of this commit is to delete some mistaken extra copies of the 'amazon' special remote, that I created in @valosekj's fork of the data-multi-subject repo, because git-annex doesn't have a command to delete a special remote; you can only create them and rename them.

The third point of this commit is to mark all the clones on various people's laptops and computers as ignored (with `git annex dead`), so that `git annex whereis` stops suggesting that annexed file be gotten from another person's computer. The only source we want to advertise for these files is amazon.

As an extra precaution for the third point, I filtered out all other mentions of the ignored clones' copies of the image files.